+++ admin/brevo-sync-page.php
@@ -11,6 +11,7 @@ function lr_render_brevo_sync_page() {
         $options = get_option('lr_brevo_options', []);
         $options['api_key'] = sanitize_text_field($_POST['brevo_api_key']);
         $options['list_folder_id'] = intval($_POST['brevo_list_folder_id']);
+        $options['sender_id'] = intval($_POST['brevo_sender_id']);
         $options['resync_days'] = intval($_POST['brevo_resync_days']);
         update_option('lr_brevo_options', $options);
         echo '<div class="notice notice-success is-dismissible"><p>Settings saved.</p></div>';
@@ -55,6 +56,13 @@ function lr_render_brevo_sync_page() {
                     <td>
                         <input type="number" id="brevo_list_folder_id" name="brevo_list_folder_id" value="<?php echo esc_attr(get_option('lr_brevo_options')['list_folder_id'] ?? 31); ?>" style="width: 100px;" />
                         <p class="description">The ID of the folder in Brevo where city lists will be created (e.g., 31 for "Cities").</p>
+                    </td>
+                </tr>
+                <tr valign="top">
+                    <th scope="row"><label for="brevo_sender_id">Brevo Sender ID</label></th>
+                    <td>
+                        <input type="number" id="brevo_sender_id" name="brevo_sender_id" value="<?php echo esc_attr(get_option('lr_brevo_options')['sender_id'] ?? ''); ?>" style="width: 100px;" />
+                        <p class="description">The numeric ID of the validated sender in your Brevo account to be used for campaigns.</p>
                     </td>
                 </tr>
                 <tr valign="top">